+++ lib/src/method_call.dart
@@ -39,21 +39,21 @@ class FlMlKitScanningMethodCall {
   /// [useEvent] 返回消息使用 FLCameraEvent
   /// The return message uses flcameraevent
   /// [rotationDegrees] Only Android is supported
-  Future<List<BarcodeModel>> scanImageByte(Uint8List uint8list,
+  Future<AnalysisImageModel?> scanImageByte(Uint8List uint8list,
       {int rotationDegrees = 0, bool useEvent = false}) async {
-    if (!_supportPlatform) return <BarcodeModel>[];
+    if (!_supportPlatform) return null;
     if (useEvent) {
       assert(
           FlCameraEvent.instance.isPaused, 'Please initialize FLCameraEvent');
     }
-    final dynamic list = await _channel.invokeMethod<dynamic>(
+    final dynamic map = await _channel.invokeMethod<dynamic>(
         'scanImageByte', <String, dynamic>{
       'byte': uint8list,
       'useEvent': useEvent,
       'rotationDegrees': rotationDegrees
     });
-    if (list != null && list is List) return getBarcodeModelList(list);
-    return <BarcodeModel>[];
+    if (map != null && map is Map) return AnalysisImageModel.fromMap(map);
+    return null;
   }
 
   /// 打开\关闭 闪光灯
@@ -65,6 +65,11 @@ class FlMlKitScanningMethodCall {
   /// Camera zoom
   Future<bool> setZoomRatio(double ratio) =>
       FlCameraMethodCall.instance.setZoomRatio(ratio);
+
+  /// 获取可用摄像头
+  /// get available Cameras
+  Future<List<CameraInfo>?> availableCameras() =>
+      FlCameraMethodCall.instance.availableCameras();
 
   /// 暂停扫描
   /// Pause scanning