+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0]
+ * add camera zoom function and add onzoomchange
+ * modify the onlisten callback value and add the width and height of the 
+   parsed picture to calculate the bar code rectangle and coordinate points displayed on the screen
+ * remove usebackcamera, add availablecameras and select the camera you need.
+   The rear camera is used by default
+ * remove zoomquality, add resolution, and set the resolution you need, 
+ * add updatereset. Do you need to reinitialize the camera preview when calling didupdatewidget
 ## [1.0.1]
  * fix the problem that there is no data in the boundingbox
 ## [1.0.0]